apps/eval-api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0
ragas==0.1.0
datasets==2.14.7
langchain-core==0.1.0
openai==1.3.0

